Solidified the Payment Link gRPC communication between Gateway, Payment and Order services.
At the moment. The Order services now successfully receives a New Order
gRPC message; Updates the new order, creating a Stripe Payment link, and
returns the order back to the Gateway and back to the front-end, which
then triggers a UI update, allowing the user to be redirected to the
Stripe Payment link.